Rename back octopus_queue.py to queue.py
I think I had changed this for some namespace collision issue,
but that does not seem to be a problem anymore so...
